chore: migrate to new openebs chart repository (#1457)
The old chart repository was archived eagerly a couple of days ago,
which broke things for a lot of people. They reinstated the repository
now, but added a deprecation notice to it.

This change switches openebs to the new chart repo. Since this does not
contain release for the 3.x branch anymore, I did the necessary values
changes to bump openebs to 4.x.

With the new values, it will only deploy the localpv-provisioner, just
like we did on version 3.x.

I'm using 4.x already since a couple of days and it's working as
intended (https://github.com/martinohmann/home-ops/blob/main/kubernetes/main/apps/openebs-system/openebs/app/helmrelease.yaml).